+++ Assets/OxGFrame/MediaFrame/Scripts/Runtime/Core/AudioFrame/AudioBase.cs
@@ -21,7 +21,7 @@ namespace OxGFrame.MediaFrame.AudioFrame
         [Tooltip("Depends on Requester.InitCacheCapacityForAudio"), ConditionalField(nameof(sourceType), true, SourceType.Audio)]
         public bool requestCached = true;
         // SourceType => StreamingAssets
-        [Tooltip("Default path is [StreamingAssets]. Just set that inside path and file name, Don't forget file name must include extension, ex: Audio/example.mp3"), ConditionalField(nameof(sourceType), false, SourceType.Streaming)]
+        [Tooltip("Default path is [StreamingAssets]. Just set that inside path and file name, Don't forget file name must with extension, ex: Audio/example.mp3"), ConditionalField(nameof(sourceType), false, SourceType.Streaming)]
         public string fullPathName = "";
         // SourceType => Url
         [ConditionalField(nameof(sourceType), false, SourceType.Url)]