+++ Assets/OxGFrame/MediaFrame/Scripts/Runtime/Core/Implement/MediaManager.cs
@@ -11,6 +11,28 @@ namespace OxGFrame.MediaFrame
         protected HashSet<string> _loadingFlags = new HashSet<string>();                                  // 用來標記正在加載中的資源 (暫存緩存)
         protected List<T> _listAllCache = new List<T>();                                                  // 【常駐】所有進入播放的影音柱列緩存 (只會在 Destroy 時, Remove 對應的緩存)
 
+        private void FixedUpdate()
+        {
+            if (this._listAllCache.Count > 0)
+            {
+                float dt1 = Time.fixedDeltaTime;
+                float dt2 = Time.fixedUnscaledDeltaTime;
+
+                int count = this._listAllCache.Count;
+                for (int i = 0; i < count; i++)
+                {
+                    if (this._listAllCache.Count != count) break;
+
+                    // 僅刷新激活的物件
+                    if (this._listAllCache[i].gameObject.activeSelf)
+                    {
+                        if (this._listAllCache[i].ignoreTimeScale) this._listAllCache[i].DriveFixedUpdate(dt2);
+                        else this._listAllCache[i].DriveFixedUpdate(dt1);
+                    }
+                }
+            }
+        }
+
         /// <summary>
         /// 檢查是否有資源緩存
         /// </summary>